Precompute the square sprites to speed up drawing
For short lived universes, this will use more memory and have a larger
startup time.
However, for longer living universes, this is faster.

Precompute all sprites, which removes the need to check if one exists
before drawing it, meaning less work needs to be done in the `on_draw`
function.